Move impl. of ank run workload into own file
Issue-Id: #229
