+++ ank/src/cli_commands.rs
@@ -26,12 +26,13 @@ mod apply_manifests;
 mod delete_workloads;
 mod get_state;
 mod get_workloads;
+mod run_workload;
 mod set_state;
 
 use common::{
     communications_error::CommunicationMiddlewareError,
     from_server_interface::FromServer,
-    objects::{CompleteState, State, StoredWorkloadSpec, Tag, WorkloadInstanceName},
+    objects::{CompleteState, State, WorkloadInstanceName},
 };
 
 use wait_list_display::WaitListDisplay;
@@ -106,47 +107,6 @@ impl CliCommands {
         }
 
         Ok(workload_infos)
-    }
-
-    // [impl->swdd~cli-provides-run-workload~1]
-    // [impl->swdd~cli-blocks-until-ankaios-server-responds-run-workload~2]
-    pub async fn run_workload(
-        &mut self,
-        workload_name: String,
-        runtime_name: String,
-        runtime_config: String,
-        agent_name: String,
-        tags_strings: Vec<(String, String)>,
-    ) -> Result<(), CliError> {
-        let tags: Vec<Tag> = tags_strings
-            .into_iter()
-            .map(|(k, v)| Tag { key: k, value: v })
-            .collect();
-
-        let new_workload = StoredWorkloadSpec {
-            agent: agent_name,
-            runtime: runtime_name,
-            tags,
-            runtime_config,
-            ..Default::default()
-        };
-        output_debug!("Request to run new workload: {:?}", new_workload);
-
-        let update_mask = vec![format!("desiredState.workloads.{}", workload_name)];
-
-        let mut complete_state_update = CompleteState::default();
-        complete_state_update
-            .desired_state
-            .workloads
-            .insert(workload_name, new_workload);
-
-        output_debug!(
-            "The complete state update: {:?}, update mask {:?}",
-            complete_state_update,
-            update_mask
-        );
-        self.update_state_and_wait_for_complete(complete_state_update, update_mask)
-            .await
     }
 
     // [impl->swdd~cli-requests-update-state-with-watch~1]
@@ -251,30 +211,22 @@ impl CliCommands {
 #[cfg(test)]
 mod tests {
     use common::{
-        commands::{UpdateStateSuccess, UpdateWorkloadState},
-        from_server_interface::{FromServer, FromServerSender},
-        objects::{
-            self, generate_test_workload_spec_with_param, CompleteState, ExecutionState, State,
-            StoredWorkloadSpec, Tag, WorkloadState,
-        },
+        from_server_interface::FromServerSender,
+        objects::{generate_test_workload_spec_with_param, CompleteState, State},
         state_manipulation::{Object, Path},
         test_utils::{self},
         to_server_interface::ToServerReceiver,
     };
-    use mockall::predicate::eq;
+
     use std::io;
 
     use super::apply_manifests::{
         create_filter_masks_from_paths, generate_state_obj_and_filter_masks_from_manifests,
         handle_agent_overwrite, parse_manifest, update_request_obj, InputSourcePair,
     };
-    use crate::{cli::ApplyArgs, cli_commands::server_connection::MockServerConnection};
+    use crate::cli::ApplyArgs;
     use serde_yaml::Value;
     use std::io::Read;
-
-    use super::CliCommands;
-
-    const RESPONSE_TIMEOUT_MS: u64 = 3000;
 
     mockall::mock! {
         pub GRPCCommunicationsClient {
@@ -285,92 +237,6 @@ mod tests {
                 agent_tx: FromServerSender,
             ) -> Result<(), String>;
         }
-    }
-
-    // [utest->swdd~cli-provides-run-workload~1]
-    // [utest->swdd~cli-blocks-until-ankaios-server-responds-run-workload~2]
-    #[tokio::test]
-    async fn utest_run_workload_one_new_workload() {
-        const TEST_WORKLOAD_NAME: &str = "name4";
-        let test_workload_agent = "agent_B".to_string();
-        let test_workload_runtime_name = "runtime2".to_string();
-        let test_workload_runtime_cfg = "some config".to_string();
-
-        let _guard = crate::test_helper::MOCKALL_CONTEXT_SYNC
-            .get_lock_async()
-            .await;
-
-        let new_workload = StoredWorkloadSpec {
-            agent: test_workload_agent.to_owned(),
-            runtime: test_workload_runtime_name.clone(),
-            tags: vec![Tag {
-                key: "key".to_string(),
-                value: "value".to_string(),
-            }],
-            runtime_config: test_workload_runtime_cfg.clone(),
-            ..Default::default()
-        };
-        let mut complete_state_update = CompleteState::default();
-        complete_state_update
-            .desired_state
-            .workloads
-            .insert(TEST_WORKLOAD_NAME.into(), new_workload);
-
-        let mut mock_server_connection = MockServerConnection::default();
-        mock_server_connection
-            .expect_update_state()
-            .with(
-                eq(complete_state_update.clone()),
-                eq(vec![format!(
-                    "desiredState.workloads.{}",
-                    TEST_WORKLOAD_NAME
-                )]),
-            )
-            .return_once(|_, _| {
-                Ok(UpdateStateSuccess {
-                    added_workloads: vec![format!(
-                        "{}.abc.agent_B",
-                        TEST_WORKLOAD_NAME.to_string()
-                    )],
-                    deleted_workloads: vec![],
-                })
-            });
-        mock_server_connection
-            .expect_get_complete_state()
-            .with(eq(vec![]))
-            .return_once(|_| Ok(Box::new(complete_state_update)));
-        mock_server_connection
-            .expect_take_missed_from_server_messages()
-            .return_once(|| {
-                vec![FromServer::UpdateWorkloadState(UpdateWorkloadState {
-                    workload_states: vec![WorkloadState {
-                        instance_name: "name4.abc.agent_B".try_into().unwrap(),
-                        execution_state: ExecutionState {
-                            state: objects::ExecutionStateEnum::Running(
-                                objects::RunningSubstate::Ok,
-                            ),
-                            additional_info: "".to_string(),
-                        },
-                    }],
-                })]
-            });
-
-        let mut cmd = CliCommands {
-            _response_timeout_ms: RESPONSE_TIMEOUT_MS,
-            no_wait: false,
-            server_connection: mock_server_connection,
-        };
-
-        let run_workload_result = cmd
-            .run_workload(
-                TEST_WORKLOAD_NAME.into(),
-                test_workload_runtime_name,
-                test_workload_runtime_cfg,
-                test_workload_agent,
-                vec![("key".to_string(), "value".to_string())],
-            )
-            .await;
-        assert!(run_workload_result.is_ok());
     }
 
     // [utest->swdd~cli-apply-supports-ankaios-manifest~1]
